requirements.txt
selenium>=4.0.0
PyPDF2>=3.0.0
Pillow>=10.0.0